+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Numbers in parentheses after entries refer to issues in the [GitHub issue tracke
 - A `detect_renaming` parameter has been added to `Database.diff` and `Database.migrate`, and a `--no-rename` flag has been added to `isqlite diff` and `isqlite migrate`, allowing isqlite's renaming detection to be disabled in case it gives erroneous results.
 
 ### Fixed
+- Previously, under some circumstances `Database.diff` would identify a column as renamed (e.g., `a --> b`) even though the old column (`a`) was still in the table. This has been fixed. (#67)
 - `isqlite create` and `isqlite update` no longer crash when a value in a command-line argument contains an equals sign.
 - `isqlite get` now prints a helpful error message instead of crashing when the row does not exist.
 - `isqlite schema <table>` now prints a helpful error message when the table does not exist.